Documentation and Norminette documentation of the code, but still left some functions norminette was fixed and files that has more than 5 functions now was splitted new validation for the Floor an Celing parameters, rgb only can be 0 to 255
