+++ Makefile
@@ -8,8 +8,9 @@ OBJ_DIR = objs
 #	====================		Files      		====================
 
 NAME = cub3d
-FILES = main.c cub3d.c file_checker.c map_checker.c \
-		utils.c error.c
+FILES = main.c cub3d.c \
+		file_checker.c map_checker.c map_checker_2.c \
+		utils.c utils_2.c utils_3.c error.c
 
 OBJ = $(FILES:.c=.o)
 TARGET = $(addprefix $(OBJ_DIR)/, $(OBJ))
@@ -99,10 +100,10 @@ re: fclean all
 .PHONY: all clean fclean re
 
 norm:
-	norminette -R CheckForbiddenSourceHeader headers/*.h srcs/*.c srcs/**/*.c
+	norminette -R CheckForbiddenSourceHeader headers/cub3d.h srcs/*.c
 
 leak: all
-	valgrind --leak-check=full --show-leak-kinds=all ./$(NAME) maps/wrong15.cub
+	valgrind --leak-check=full --show-leak-kinds=all ./$(NAME) maps/valid/map.cub
 
 display:
 	@clear
