+++ maps/valid/triangle_multi_open.cub
